requirements.txt
Pillow>=9.0.0     # 用于截图功能
pyyaml>=6.0       # 用于配置文件处理
schedule>=1.1.0   # 用于任务调度